tidy and form files into same structure as content webapp for prismic
