+++ catalogue/webapp/pages/search/stories.tsx
@@ -5,7 +5,7 @@ import { getServerData } from '@weco/common/server-data';
 import Space from '@weco/common/views/components/styled/Space';
 import { NextPageWithLayout } from '@weco/common/views/pages/_app';
 import { getSearchLayout } from 'components/SearchPageLayout/SearchPageLayout';
-import { getStories } from '../../services/prismic';
+import { getStories } from '../../services/prismic/fetch/articles';
 
 type Props = {
   storyResponseList;
